# shared/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shravya API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/teacher/courses/{course_id}/contexts": {
      "get": {
        "tags": [
          "curriculum"
        ],
        "summary": "List Contexts",
        "operationId": "list_teacher_contexts",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_list_ContextSummaryResponse__"
                }
              }
            }
          },
          "403": {
            "description": "The requested action is not permitted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested curriculum record was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The curriculum context is in a conflicting state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The curriculum request could not be completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teacher/contexts/{context_id}/completeness": {
      "get": {
        "tags": [
          "curriculum"
        ],
        "summary": "Get Completeness Route",
        "operationId": "get_context_completeness",
        "parameters": [
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Context Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_CompletenessResponse_"
                }
              }
            }
          },
          "403": {
            "description": "The requested action is not permitted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested curriculum record was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The curriculum context is in a conflicting state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The curriculum request could not be completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teacher/contexts/{context_id}/review-events": {
      "get": {
        "tags": [
          "curriculum"
        ],
        "summary": "History",
        "operationId": "get_review_events",
        "parameters": [
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Context Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_list_ReviewEventResponse__"
                }
              }
            }
          },
          "403": {
            "description": "The requested action is not permitted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested curriculum record was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The curriculum context is in a conflicting state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The curriculum request could not be completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teacher/contexts/{context_id}": {
      "get": {
        "tags": [
          "curriculum"
        ],
        "summary": "Detail",
        "operationId": "get_teacher_context",
        "parameters": [
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Context Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_ContextDetailResponse_"
                }
              }
            }
          },
          "403": {
            "description": "The requested action is not permitted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested curriculum record was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The curriculum context is in a conflicting state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The curriculum request could not be completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teacher/contexts/{context_id}/submit-for-review": {
      "post": {
        "tags": [
          "curriculum"
        ],
        "summary": "Submit",
        "operationId": "submit_context_review",
        "parameters": [
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Context Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReviewNoteRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_SubmitResponse_"
                }
              }
            }
          },
          "403": {
            "description": "The requested action is not permitted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested curriculum record was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The curriculum context is in a conflicting state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The curriculum request could not be completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teacher/contexts/{context_id}/return-to-draft": {
      "post": {
        "tags": [
          "curriculum"
        ],
        "summary": "Return Draft",
        "operationId": "return_context_draft",
        "parameters": [
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Context Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReviewNoteRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_ContextSummaryResponse_"
                }
              }
            }
          },
          "403": {
            "description": "The requested action is not permitted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested curriculum record was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The curriculum context is in a conflicting state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The curriculum request could not be completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teacher/contexts/{context_id}/approve": {
      "post": {
        "tags": [
          "curriculum"
        ],
        "summary": "Approve",
        "operationId": "approve_context",
        "parameters": [
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Context Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_ApprovalResponse_"
                }
              }
            }
          },
          "403": {
            "description": "The requested action is not permitted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested curriculum record was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The curriculum context is in a conflicting state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The curriculum request could not be completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teacher/contexts/{context_id}/copy-to-new-draft": {
      "post": {
        "tags": [
          "curriculum"
        ],
        "summary": "Copy",
        "operationId": "copy_context_draft",
        "parameters": [
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Context Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CopyRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_ContextSummaryResponse_"
                }
              }
            }
          },
          "403": {
            "description": "The requested action is not permitted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested curriculum record was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The curriculum context is in a conflicting state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The curriculum request could not be completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "error",
            "title": "Status",
            "default": "error"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "message_key": {
            "type": "string",
            "title": "Message Key"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          },
          "recoverable": {
            "type": "boolean",
            "title": "Recoverable"
          },
          "next_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Next Actions"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "message_key",
          "recoverable"
        ],
        "title": "ErrorResponse"
      },
...
      "ReviewNoteRequest": {
        "properties": {
          "reviewer_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Note"
          }
        },
        "type": "object",
        "title": "ReviewNoteRequest"
      },
...
      "SuccessResponse_CompletenessResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/CompletenessResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[CompletenessResponse]"
      },
      "SuccessResponse_ContextDetailResponse_": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/ContextDetailResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[ContextDetailResponse]"
      },
...
      "SuccessResponse_list_ContextSummaryResponse__": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "default": "success"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ContextSummaryResponse"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[list[ContextSummaryResponse]]"
      },
      "SuccessResponse_list_ReviewEventResponse__": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "default": "success"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ReviewEventResponse"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[list[ReviewEventResponse]]"
      },
...
    }
  }
}